Moved algorithm traits to algorithms
