+++ src/dachshund/algorithms/adjacency_matrix.rs
