+++ src/dachshund/algorithms/mod.rs
@@ -4,5 +4,15 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
+pub mod adjacency_matrix;
+pub mod algebraic_connectivity;
+pub mod betweenness;
+pub mod clustering;
+pub mod cnm_communities;
 pub mod connected_components;
+pub mod connectivity;
 pub mod coreness;
+pub mod eigenvector_centrality;
+pub mod laplacian;
+pub mod shortest_paths;
+pub mod transitivity;
