+++ src/dachshund/mod.rs
@@ -4,22 +4,14 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-pub mod adjacency_matrix;
-pub mod algebraic_connectivity;
 pub mod algorithms;
 pub mod beam;
-pub mod betweenness;
 pub mod candidate;
-pub mod clustering;
-pub mod cnm_communities;
-pub mod connectivity;
-pub mod eigenvector_centrality;
 pub mod error;
 pub mod graph_base;
 pub mod graph_builder;
 pub mod id_types;
 pub mod input;
-pub mod laplacian;
 pub mod line_processor;
 pub mod node;
 pub mod non_core_type_ids;
@@ -27,14 +19,12 @@ pub mod output;
 pub mod row;
 pub mod scorer;
 pub mod search_problem;
-pub mod shortest_paths;
 pub mod simple_transformer;
 pub mod simple_undirected_graph;
 pub mod simple_undirected_graph_builder;
 pub mod test_utils;
 pub mod transformer;
 pub mod transformer_base;
-pub mod transitivity;
 pub mod typed_graph;
 pub mod typed_graph_builder;
 pub mod typed_graph_line_processor;
