+++ tests/triangles.rs
@@ -9,11 +9,11 @@
 extern crate lib_dachshund;
 extern crate test;
 
-use lib_dachshund::dachshund::clustering::Clustering;
+use lib_dachshund::dachshund::algorithms::clustering::Clustering;
 use lib_dachshund::dachshund::id_types::NodeId;
 use lib_dachshund::dachshund::simple_undirected_graph::SimpleUndirectedGraph;
 use lib_dachshund::dachshund::simple_undirected_graph_builder::SimpleUndirectedGraphBuilder;
-use lib_dachshund::dachshund::transitivity::Transitivity;
+use lib_dachshund::dachshund::algorithms::transitivity::Transitivity;
 
 use test::Bencher;
